+++ src/main/kotlin/com/wutsi/codegen/kotlin/server/ServerCodeGenerator.kt
@@ -24,7 +24,8 @@ class ServerCodeGenerator(
         MQueueCodeGenerator(),
         SwaggerCodeGenerator(),
         ServerReadmeCodeGenerator(),
-        CorsCodeGenerator()
+        CorsCodeGenerator(),
+        TracingCodeGenerator()
     )
 ) : CodeGenerator {