+++ src/test/kotlin/com/wutsi/codegen/kotlin/sdk/SdkApiBuilderCodeGeneratorTest.kt
@@ -32,13 +32,21 @@ internal class SdkApiBuilderCodeGeneratorTest {
                 package com.wutsi.test
 
                 import com.fasterxml.jackson.databind.ObjectMapper
+                import feign.RequestInterceptor
+                import kotlin.collections.List
 
                 public class TestApiBuilder {
-                  public fun build(env: Environment, mapper: ObjectMapper) = feign.Feign.builder()
+                  public fun build(
+                    env: Environment,
+                    mapper: ObjectMapper,
+                    interceptors: List<RequestInterceptor> = emptyList()
+                  ) = feign.Feign.builder()
                     .client(feign.okhttp.OkHttpClient())
                     .encoder(feign.jackson.JacksonEncoder(mapper))
                     .decoder(feign.jackson.JacksonDecoder(mapper))
-                    .logger(feign.slf4j.Slf4jLogger())
+                    .logger(feign.slf4j.Slf4jLogger(TestApi::class.java))
+                    .logLevel(feign.Logger.Level.BASIC)
+                    .requestInterceptors(interceptors)
                     .target(TestApi::class.java, env.url)
                 }
             """.trimIndent(),