+++ src/test/kotlin/com/wutsi/codegen/kotlin/server/ServerCodeGeneratorTest.kt
@@ -37,7 +37,7 @@ internal class ServerCodeGeneratorTest {
     fun `generators`() {
         val codegen = ServerCodeGenerator(mapper = KotlinMapper(context))
 
-        assertEquals(15, codegen.generators.size)
+        assertEquals(16, codegen.generators.size)
         assertTrue(codegen.generators[0] is ServerDtoCodeGenerator)
         assertTrue(codegen.generators[1] is ServerDelegateCodeGenerator)
         assertTrue(codegen.generators[2] is ServerControllerCodeGenerator)
@@ -53,5 +53,6 @@ internal class ServerCodeGeneratorTest {
         assertTrue(codegen.generators[12] is SwaggerCodeGenerator)
         assertTrue(codegen.generators[13] is ServerReadmeCodeGenerator)
         assertTrue(codegen.generators[14] is CorsCodeGenerator)
+        assertTrue(codegen.generators[15] is TracingCodeGenerator)
     }
 }