+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # See https://keepachangelog.com/en/1.0.0/
 
-## [0.0.26] 2021-04-08
+## [0.0.27] 2021-04-10
+### Added
+- Add request interceptors
+- slf4j configuration
+### Fixed
+- Default value for `Boolean`
+
+## [0.0.26] 2021-04-09
 ### Fixed
 - Return value of Sdk APIBuilder